// WebApplicationPlateforme/ClientApp/src/app/shared/Models/MediaCenter/film/film.model.ts
export class Film {

  id: number;
  typefilm: string;
  autretype: string;
  budget: string;
  duree: string;
  but: string;
  url: string;
  dirid: string;
  dirnom: string;
  diretat: string;
  dirdate: string;
  adminid: string;
  adminom: string;
  admietat: string;
  admidate: string;
  etabnom: string;
  etabid: string;
  etabetat: string;
  etabdate: string;
  attribut1: string;
  attribut2: string;
  attribut3: string;
  attribut4: string;
  attribut5: string;
  attribut6: string;
  dateenreg: string;
  userNameCreator: string;
  idUserCreator: string;
}